requirements.txt
pygetwindow
pyautogui
Pillow
pyinstaller
